Refactor setup models into PPO strategy dataclasses
