+++ crypto_screener/domain/models/setups/__init__.py
@@ -1,0 +1,8 @@
+from crypto_screener.domain.models.setups.ppo import Capture, Setup, Trade, Unfilled
+
+__all__ = [
+    "Capture",
+    "Setup",
+    "Trade",
+    "Unfilled",
+]
