+++ crypto_screener/execution/test_result.py
@@ -2,7 +2,7 @@ from collections import defaultdict
 from typing import Optional
 
 from crypto_screener.domain.models.bar import Bar
-from crypto_screener.domain.models.setup import Buy
+from crypto_screener.domain.models.setups.ppo import Trade
 from crypto_screener.domain.models.trade_result import TradeResult
 from crypto_screener.utils.logger import log
 
@@ -19,7 +19,7 @@ def _get_profit_pct(
 # endregion
 
 def evaluate_buy(
-        setup: Buy,
+        setup: Trade,
         future_bars: list[Bar]
 ) -> Optional[tuple[TradeResult, float]]:
     entry_price = setup.entry_price
